+++ Editor/SimpleSettings.cs
@@ -1,0 +1,14 @@
+﻿using UnityEditor;
+using UnityEngine;
+
+namespace SubstanceToTexture {
+	public class SimpleSettings<T> : ScriptableSingleton<T> where T : SimpleSettings<T> {
+		public void OnEnable() {
+			hideFlags &= ~HideFlags.NotEditable;
+		}
+
+		public void Save() {
+			base.Save(true);
+		}
+	}
+}